+++ CondorUIComponents/UIModels/Selectable.swift
@@ -1,0 +1,11 @@
+//
+//  Selectable.swift
+//  CondorUIComponents
+//
+//  Created by Luis David Goyes Garces on 3/6/19.
+//  Copyright © 2019 Condor Labs. All rights reserved.
+//
+
+public protocol Selectable {
+    func getSelectableText() -> String
+}
