+++ CondorUIComponents/UIModels/Text/Date/DateFormFieldChangeNotifiable.swift
@@ -6,12 +6,12 @@
 //  Copyright © 2019 Condor Labs. All rights reserved.
 //
 
-protocol DateFormFieldChangeNotifiable: class {
+public protocol DateFormFieldChangeNotifiable: class {
     func onSelected(dateString: String, date: Date, from datePicker: DateFormField)
     func onDoneButtonPressed(from datePicker: DateFormField)
 }
 
-extension DateFormFieldChangeNotifiable {
+public extension DateFormFieldChangeNotifiable {
     func onDoneButtonPressed(from datePicker: DateFormField) {
         // Method intentionally left in blank
     }
