+++ CondorUIComponents/UIModels/Text/Regex/TextRegexFormField.swift
@@ -10,47 +10,32 @@ import UIKit
 
 public protocol TextRegexFormFieldProtocol: TextFormFieldType {
     func set(regex: String)
+    func getRegex() -> String
 }
 
 public class TextRegexFormField: TextFormField, TextRegexFormFieldProtocol {
-    
-    private var regex: String?
-    
+
+    private var regex: String = Constants.Regex.any
+
     public func set(regex: String) {
         self.regex = regex
     }
-    
+
+    public func getRegex() -> String {
+        return self.regex
+    }
+
     override func validateContent() -> ValidationResult {
         guard let text = self.getValue(), !text.isEmpty else {
             return ValidationResult(isValid: false, error: FormFieldError.emptyField)
         }
-        
-        guard let regexString = self.regex else {
-            return ValidationResult(isValid: false, error: TextRegexFormFieldError.missingRegex)
-        }
-        
-        let regex = NSRegularExpression(regexString)
-        
+
+        let regex = NSRegularExpression(self.regex)
+
         guard regex.matches(text) else {
             return ValidationResult(isValid: false, error: TextRegexFormFieldError.notMatchingRegex)
         }
-        
+
         return ValidationResult(isValid: true)
-    }
-}
-
-public enum TextRegexFormFieldError: FormFieldErrorType {
-    case missingRegex
-    case notMatchingRegex
-}
-
-extension TextRegexFormFieldError {
-    public var description: String? {
-        switch self {
-        case .missingRegex:
-            return "Missing regex."
-        case .notMatchingRegex:
-            return "Regex does not match."
-        }
     }
 }
