+++ CondorUIComponents/UIModels/Text/Regex/TextRegexFormFieldError.swift
@@ -1,0 +1,23 @@
+//
+//  TextRegexFormFieldError.swift
+//  CondorUIComponents
+//
+//  Created by Luis David Goyes Garces on 3/6/19.
+//  Copyright © 2019 Condor Labs. All rights reserved.
+//
+
+public enum TextRegexFormFieldError: FormFieldErrorType {
+    case missingRegex
+    case notMatchingRegex
+}
+
+extension TextRegexFormFieldError {
+    public var description: String? {
+        switch self {
+        case .missingRegex:
+            return "Missing regex."
+        case .notMatchingRegex:
+            return "Regex does not match."
+        }
+    }
+}
